Se agregó documentación al proyecto

+++ src/Controller/Cliente/DeleteController.php
@@ -6,7 +6,13 @@ use App\Repository\ClienteRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
+/**
+ * Representa eliminación de datos en el servidor
+ * HTTP Method: DELETE
+ * HTTP Status Code: 204, 404
+ */
 class DeleteController extends AbstractController
 {
     /** @required */
@@ -15,6 +21,10 @@ class DeleteController extends AbstractController
     public function __invoke(Request $request, string $id): Response
     {
         $cliente = $this->clienteRepository->find($id);
+
+        if (!$cliente) {
+            throw new NotFoundHttpException();
+        }
 
         if ($this->isCsrfTokenValid('delete'.$cliente->getId(), $request->request->get('_token'))) {
             $this->clienteRepository->remove($cliente, true);
